# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Deep Control Loads Backend Rest API Documentation",
        "description": "Rest API Documentation",
        "contact": {
            "email": "remstroyod@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/orders/{order}/documents/{document}": {
            "delete": {
                "tags": [
                    "Admin: Orders"
                ],
                "summary": "Delete Document Rest API",
                "description": "Delete Document",
                "operationId": "admin.orders.documents.destroy",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID Order",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "document",
                        "in": "path",
                        "description": "ID Document",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/admin/users/{user}": {
            "put": {
                "tags": [
                    "Admin: Users"
                ],
                "summary": "Update User Rest API",
                "description": "Update User",
                "operationId": "admin.users.update",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "User Status",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "role": {
                                        "description": "User Role",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "users": {
                                        "description": "Users Attach Manager",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin: Users"
                ],
                "summary": "Delete User Rest API",
                "description": "Delete User",
                "operationId": "admin.users.destroy",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/drivers/{driver}": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Show Driver Rest API",
                "description": "Show Driver",
                "operationId": "drivers.show",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "description": "ID Driver",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Update Driver Rest API",
                "description": "Update Driver",
                "operationId": "drivers.update",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "description": "ID Driver",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "gender": {
                                        "description": "Gender",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "first_name": {
                                        "description": "First Name",
                                        "type": "string",
                                        "example": "Driver Name"
                                    },
                                    "surname": {
                                        "description": "Surname",
                                        "type": "string",
                                        "example": "Drive Surname"
                                    },
                                    "phone": {
                                        "description": "Phone",
                                        "type": "string",
                                        "example": "+380960000000"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "example": "example@hostname.com"
                                    },
                                    "has_license": {
                                        "description": "License",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "languages": {
                                        "description": "Languages",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "RU",
                                            "BG",
                                            "PL"
                                        ]
                                    },
                                    "file": {
                                        "description": "Upload File",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_type": {
                                        "description": "Type File",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "file_number": {
                                        "description": "Number File",
                                        "type": "string",
                                        "example": "12-RT-54"
                                    },
                                    "file_valid_from": {
                                        "description": "Date Valid From File",
                                        "type": "string",
                                        "example": "2024-06-28"
                                    },
                                    "file_valid_until": {
                                        "description": "Date Valid Until File",
                                        "type": "string",
                                        "example": "2024-06-28"
                                    },
                                    "_method": {
                                        "description": "Method (should always be PUT)",
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Delete Driver Rest API",
                "description": "Delete Driver",
                "operationId": "drivers.destroy",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "description": "ID Driver",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/orders/{order}/documents/{document}": {
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete Document Rest API",
                "description": "Delete Document",
                "operationId": "orders.documents.destroy",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID Order",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "document",
                        "in": "path",
                        "description": "ID Document",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/results": {
            "get": {
                "tags": [
                    "Results"
                ],
                "summary": "Get Results Rest API",
                "description": "Get Results",
                "operationId": "results.index",
                "requestBody": {
                    "description": "Create user object",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "dateStart": {
                                        "description": "Date Start",
                                        "type": "string",
                                        "example": "2024-05-26"
                                    },
                                    "dateEnd": {
                                        "description": "Date End",
                                        "type": "string",
                                        "example": "2024-05-26"
                                    },
                                    "countryStart": {
                                        "description": "Country Start",
                                        "type": "string",
                                        "example": "GR"
                                    },
                                    "countryEnd": {
                                        "description": "Country End",
                                        "type": "string",
                                        "example": "BG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/results/{id}": {
            "get": {
                "tags": [
                    "Results"
                ],
                "summary": "Get Result Offer Rest API",
                "description": "Get Result Offer",
                "operationId": "results.index.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Document",
                        "required": true,
                        "example": "b5380f06-7a54-f78d-aa4e-8a3b79b1c3c6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}